Add Network Manager
