+++ azure-communication-ui/azure-communication-ui/src/calling/java/com/azure/android/communication/ui/calling/di/DependencyInjectionContainerImpl.kt
@@ -19,6 +19,7 @@ import com.azure.android.communication.ui.calling.presentation.manager.CameraSta
 import com.azure.android.communication.ui.calling.presentation.manager.LifecycleManagerImpl
 import com.azure.android.communication.ui.calling.presentation.manager.MeetingJoinedHook
 import com.azure.android.communication.ui.calling.presentation.manager.MicStatusHook
+import com.azure.android.communication.ui.calling.presentation.manager.NetworkManager
 import com.azure.android.communication.ui.calling.presentation.manager.ParticipantAddedOrRemovedHook
 import com.azure.android.communication.ui.calling.presentation.manager.PermissionManager
 import com.azure.android.communication.ui.calling.presentation.manager.SwitchCameraStatusHook
@@ -96,6 +97,12 @@ internal class DependencyInjectionContainerImpl(
     override val audioFocusManager by lazy {
         AudioFocusManager(
             appStore,
+            applicationContext,
+        )
+    }
+
+    override val networkManager by lazy {
+        NetworkManager(
             applicationContext,
         )
     }
